refactor: 고정 지출 관련 피드백 적용

+++ server/back/src/main/java/com/barcode/honeykeep/fixedexpense/repository/FixedExpenseRepository.java
@@ -7,6 +7,6 @@ import java.util.List;
 
 public interface FixedExpenseRepository extends JpaRepository<FixedExpense, Long> {
 
-    List<FixedExpense> findByUserId(Integer userId);
+    List<FixedExpense> findByUserId(Long userId);
 
 }
